+++ backend/ezgloszenie/src/main/java/io/github/pluton33/ezgloszenie/data/User.java
@@ -7,7 +7,7 @@ import java.time.LocalDate;
 
 
 public record User(
-        @Nullable Integer  id,
+        @Nullable Long  id,
         UserRole role,
         String email,
         Integer badgeNumber,
