+++ backend/ezgloszenie/src/main/java/io/github/pluton33/ezgloszenie/repository/ReportsRepository.java
@@ -3,5 +3,5 @@ package io.github.pluton33.ezgloszenie.repository;
 import io.github.pluton33.ezgloszenie.data.ReportEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ReportsRepository extends JpaRepository<ReportEntity, Integer> {
+public interface ReportsRepository extends JpaRepository<ReportEntity, Long> {
 }
