+++ backend/ezgloszenie/src/main/java/io/github/pluton33/ezgloszenie/repository/UsersRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UsersRepository extends JpaRepository <UserEntity, Integer> {
+public interface UsersRepository extends JpaRepository <UserEntity, Long> {
     Optional<UserEntity> findByEmail(String email);
 
     boolean existsByEmail(String email);
